Voice: pitch calculations are doubles.

+++ Voice.c
@@ -69,7 +69,7 @@ void Voice_render(Voice* self, uint32_t num_frames, float* l_out, float* r_out)
 
 	double tet = Voice_param_value(self, TET_PARAM);
 	double a_hz = Voice_param_value(self, A_HZ_PARAM);
-	float phase_increment = a_hz * exp2f((self->key - 57) / tet) / self->plugin->sample_rate;
+	double phase_increment = a_hz * exp2((self->key - 57) / tet) / self->plugin->sample_rate;
 	float gain_change_per_sample = 0.0;
 	if (self->state == ATTACK)
 		gain_change_per_sample = 1.0 / (self->plugin->sample_rate * Voice_param_value(self, ATTACK_PARAM));
